--- a/WebDriverAgentLib/Categories/XCUIElement+FBUID.m
+++ b/WebDriverAgentLib/Categories/XCUIElement+FBUID.m
@@ -12,6 +12,7 @@
 #import "FBElementUtils.h"
 #import "XCUIApplication.h"
 #import "XCUIElement+FBUtilities.h"
+#import <objc/runtime.h>
 
 @implementation XCUIElement (FBUID)
 
@@ -33,6 +34,19 @@
 
 @implementation FBXCElementSnapshotWrapper (FBUID)
 
+#pragma clang diagnostic push
+#pragma clang diagnostic ignored "-Wobjc-load-method"
++ (void)load
+{
+  Class XCElementSnapshotCls = objc_lookUpClass("XCElementSnapshot");
+  if (XCElementSnapshotCls != nil)
+  {
+    Method uidMethod = class_getInstanceMethod(self.class, @selector(fb_uid));
+    class_addMethod(XCElementSnapshotCls, @selector(fb_uid), method_getImplementation(uidMethod), method_getTypeEncoding(uidMethod));
+  }
+}
+#pragma diagnostic pop
+
 - (unsigned long long)fb_accessibiltyId
 {
   return [FBElementUtils idWithAccessibilityElement:self.accessibilityElement];
@@ -45,7 +59,10 @@
 
 - (NSString *)fb_uid
 {
-  return [self.class wdUIDWithSnapshot:self.snapshot];
+  if ([self isKindOfClass:FBXCElementSnapshotWrapper.class]) {
+    return [self.class wdUIDWithSnapshot:self.snapshot];
+  }
+  return [FBElementUtils uidWithAccessibilityElement:[self accessibilityElement]];
 }
 
 @end