@mykola-mokhnach 2023-02-22
IMHO it is never a good idea to use XPath location unless there is absolutely no other way to find an element (e.g. axes or special XPathfunctions are needed)