@Dan-Maor 2023-02-23
Completely agree that Xpath is the worst option as well.