@hdesai-dave 2023-03-09
Can we backport this change to appium 1.22.x since appium 2.0 is still in beta?