@jlipps 2023-03-10
@hdesai-dave Appium 2 is the only supported version of Appium. It's only in beta since the documentation is not yet fully fully complete (but it's 99% there). For all intents and purposes Appium 2 is out of beta and should be the only version used going forward. All that to say, we won't be backporting anything to 1.x.