@Jon889 2023-03-23
Does this, block predicates having worse performance, mean that the usage of block predicates for iOS predicates sent from the appium client also have a negative performance impact? https://github.com/appium/WebDriverAgent/blob/209d01a680003fd4864061487b1c3a4e0b76b2db/WebDriverAgentLib/Categories/XCUIElement%2BFBFind.m#L80